add screenshot tests
